# backend/openapi.yaml
openapi: 3.0.0
info:
  version: 2.0.1
  title: Anki Clone Application API
  description: This is the API of the Anki Clone Application. It allows you to manage users, learning packages, learning facts, tags and statistics.
paths:
...
  /api/user/{user_id}:
    get:
      summary: Get a user by id
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update a user by id
      operationId: updateUserById
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user to update
          required: true
          schema:
            type: number
      requestBody:
        description: The user to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                  description: The mail of the user
                  example: john.doe@gmail.com
                pseudo:
                  type: string
                  description: The pseudo of the user
                  example: John
                password:
                  type: string
                  description: The password of the user
                  example: 123456
                profilePicture:
                  type: string
                  description: The profile picture of the user
                  example: https://www.google.com/
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /api/fact:
    get:
      summary: Get all learning facts
      operationId: getLearningFact
      tags:
        - Learning Facts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningFact'
        '500':
          description: Internal Server Error
  /api/fact/{fact_id}:
    get:
      summary: Get a learning fact by id
      operationId: getLearningFactById
      tags:
        - Learning Facts
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningFact'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      summary: Update a learning fact by id
      operationId: updateLearningFactById
      tags:
        - Learning Facts
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
      requestBody:
        description: The learning fact to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                front:
                  type: string
                  description: The front of the learning fact
                  example: What is TypeScript?
                back:
                  type: string
                  description: The back of the learning fact
                  example: TypeScript is a statically typed superset of JavaScript.
                source:
                  type: string
                  description: The source of the learning fact
                  example: https://www.google.com/
                relatedImage:
                  type: string
                  description: The related image of the learning fact
                  example: https://www.google.com/
                relatedLink:
                  type: string
                  description: The related link of the learning fact
                  example: https://www.google.com/
                packageId:
                  type: number
                  description: The id of the learning package
                  example: 1
                creatorId:
                  type: number
                  description: The id of the creator of the learning fact
                  example: 1
      responses:
        '204':
          description: Updated
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a learning fact by id
      operationId: deleteLearningFactById
      tags:
        - Learning Facts
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Deleted
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/fact/{fact_id}/deactivate:
    put:
      summary: Deactivate a learning fact by id
      operationId: deactivateLearningFactById
      tags:
        - Learning Facts
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
      responses:
        '204':
          description: Updated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # TAG API
...
  /api/package/{package_id}/fact:
    get:
      summary: Get all learning facts by learning package id
      operationId: getLearningFactByLearningPackageId
      tags:
        - Learning Packages <> Learning Facts
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearningFact'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new learning fact
      operationId: createLearningFact
      tags:
        - Learning Packages <> Learning Facts
      parameters:
        - name: package_id
          in: path
          description: The id of the learning package
          required: true
          schema:
            type: number
      requestBody:
        description: The learning fact to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                front:
                  type: string
                  description: The front of the learning fact
                  example: What is TypeScript?
                  required: true
                back:
                  type: string
                  description: The back of the learning fact
                  example: TypeScript is a statically typed superset of JavaScript.
                  required: true
                source:
                  type: string
                  description: The source of the learning fact
                  example: https://www.google.com/
                  required: false
                relatedImage:
                  type: string
                  description: The related image of the learning fact
                  example: https://www.google.com/
                  required: false
                relatedLink:
                  type: string
                  description: The related link of the learning fact
                  example: https://www.google.com/
                  required: false
                creatorId:
                  type: number
                  description: The id of the creator of the learning fact
                  example: 1
                  required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningFact'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '500':
          description: Internal Server Error

  # LEARNING PACKAGE <> TAG API
...
  /api/fact/user/{user_id}:
    get:
      summary: Get all learning facts by user id
      operationId: getLearningFactByUserId
      tags:
        - Learning Facts <> Users
      parameters:
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A list of learning facts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserLearningFact'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/fact/{fact_id}/user/{user_id}:
    get:
      summary: Get a learning fact by id and user id
      operationId: getLearningFactByIdAndUserId
      tags:
        - Learning Facts <> Users
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      responses:
        '200':
          description: A learning fact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLearningFact'
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /api/fact/{fact_id}/review/{user_id}:
    put:
      summary: Review a learning fact by id and user id
      operationId: reviewLearningFactByIdAndUserId
      tags:
        - Learning Facts <> Users
      parameters:
        - name: fact_id
          in: path
          description: The id of the learning fact
          required: true
          schema:
            type: number
        - name: user_id
          in: path
          description: The id of the user
          required: true
          schema:
            type: number
      requestBody:
        description: The user learning fact to update
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confidenceLevel:
                  type: string
                  description: The confidence level of the user learning fact
                  example: 3
      responses:
        '204':
          description: Updated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
servers:
  [
    {
      url: http://localhost:3000,
      description: Development server
    }
  ]
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: number
          description: The id of the user
          example: 1
        mail:
          type: string
          description: The mail of the user
          example: john.doe@gmail.com
        pseudo:
          type: string
          description: The pseudo of the user
          example: John
        password:
          type: string
          description: The password of the user
          example: 123456
        registrationDate:
          type: string
          description: The registration date of the user
          example: 2021-01-01
        profilePicture:
          type: string
          description: The profile picture of the user
          example: https://www.google.com/
        isActive:
          type: boolean
          description: The status of the user
          example: true
      required:
        - userId
        - mail
        - pseudo
        - password
        - registrationDate
        - isActive
...
    LearningFact:
      type: object
      properties:
        factId:
          type: number
          description: The id of the learning fact
          example: 1
        front:
          type: string
          description: The front of the learning fact
          example: What is TypeScript?
        back:
          type: string
          description: The back of the learning fact
          example: TypeScript is a statically typed superset of JavaScript.
        source:
          type: string
          description: The source of the learning fact
          example: https://www.google.com/
        relatedImage:
          type: string
          description: The related image of the learning fact
          example: https://www.google.com/
        relatedLink:
          type: string
          description: The related link of the learning fact
          example: https://www.google.com/
        isActive:
          type: boolean
          description: The status of the learning fact
          example: true
        packageId:
          type: number
          description: The id of the learning package
          example: 1
        creatorId:
          type: number
          description: The id of the creator of the learning fact
          example: 1
      required:
        - factId
        - front
        - back
        - source
        - isActive
        - packageId
        - creatorId
...
    UserLearningFact:
      type: object
      properties:
        userLearningFactId:
          type: number
          description: The id of the user learning fact
          example: 1
        reviewCount:
          type: number
          description: The review count of the user learning fact
          example: 1
        confidenceLevel:
          type: string
          description: The confidence level of the user learning fact
          example: 1
        lastReviewed:
          type: string
          description: The last reviewed date of the user learning fact
          example: 2021-01-01
        nextReviewDate:
          type: string
          description: The next review date of the user learning fact
          example: 2021-01-01
        factId:
          type: number
          description: The id of the learning fact
          example: 1
        userId:
          type: number
          description: The id of the user
          example: 1
      required:
        - userLearningFactId
        - reviewCount
        - confidenceLevel
        - lastReviewed
        - nextReviewDate
        - factId
        - userId
...